Phase 1: FrameWork - Feature: Thrusters
Player class now implements thrusters feature. Speed is added to the current speed while the left shift key is being pressed.

Also, the boost powerup was modified a bit so it acts similar to how thruster speed is handled.

The values of Thruster Speed and Boost Speed were changed in the prefab and scene