Phase 1: Framework - Feature: Shield Strength
Now shield resist a number of hits. The quantity of hits depend on the quantity of colors added to the new Color[] variable name _shieldStrengthColors. Each added color adds a resist hit to the shield, and the current hits left in the shield is shown with the colors in the array.