Phase 1: Framework - Feature: Thruster Scaling Bar HUD
A new element was added to the canvas to see boost usage. Boost is active when pressing Left Shift down. While boost is used, the HUD will indicate "POWER ON" and a counter will measure how many seconds the boost has been used. If the player stops using the boost before reaching the limit (5 seconds), the boost will start to decrease the time counter. In this state the player is able to activate the booster again. However, if the player reaches the limit of the boost, the HUD will change to "COOLDOWN", and the player won't be able to use the boost until the boost time reaches to zero.