Phase 1: Framework - Feature: Camera Shake
A new script "CameraShake" was added to add a little shake to the camera this script is attached too. The shake is activated by the player when it hits an enemy. The strength and duration of the shake effect can be tuned.

Other:
Also, moved BoostHUD Script to Assets/Scripts