Phase II: Core Programming - Feature:New Enemy Movement
Movement is now generated by classes. All this movement classes implement the interface IMovement, which only defines a function to obtain a movement vector. The used class defines the result of the vector patters obtained. Right now we have two movement types. One is Straight down, and the other is side to side while moving down (HorizontalWaveDownMovement).

This movements were applied to the Enemy class, and can be changed with the function ChangeMovementTypeTo. This classes can be used by any other class that needs a movement pattern, and more patterns can be added later.

The SpawnManager was modified to set a random movement type to each instantiated enemy