Phase II: Core Programming - Feature: Wave System
Game now has Levels, each level decreases spawn time of the enemies, and increases the number of enemies spawned at a time or the level duration.

WaveCreator
A new class that contains the tools needed to create a wave template, that later will be used by the Spawn Manager.

WaveData (WaveCreator.cs)
The class that contains the properties of a wave. It's the wave template. It also delivers a random (and random with weights) pickup or enemy.

GameObjectSpawnData (WaveCreator.cs)
Contains the reference of an enemy or pickup as a gGmeObject, along with some properties.

Range (WaveCreator.cs)
Contains the definition of a range (max and min value)

WaveDirector
It is the class that defines the properties of each wave created, and delivers this info to the caller class.

UIManager
It now contains a new Text to show the intro to the level which shows as "Level X"

Spawn Manager
The class is no longer in charge of handling the enemy references, pickup references, or calculate the random selection. However, now is in charge of managing the level progression, requesting waves and working with the data delivered to know how much enemies must initialize at once, how much time the wave will run, and when to continue with the next wave. Also, is the one how calls the UIManager to show the text of the Level number.