Phase II: Core Programming - Preparations for New Enemy Types
Enemy behaviour has been modularized. For that I created the scripts:
- EnemyCore: Enemy state, life, and collisions (as properties grow, enemy core will be modularized in more parts, for example, life managers or collision manager)
- EnemyMovement: The standard movement of going straight down. This script can be interchanged with future behaviours for movement.
- EnemyStandardWeapon: The basic weapon behaviour. Later we can add more weapon behaviours.

With the scripts organized like this, I can interchange pieces of behaviour (which will be always looking at EnemyCore to know when they must be active or not).

I added a new Enemy prefab called SharpShooter. Currently behaves the same as the initial enemy, but now is composed of core, movement and weapon scripts. In the next commits i will create another movement and weapon type.

A new type of laser is added, but is till in progress (SharpshootLaser).

A new type of IMovement was added, NoMovement, only used if the ship will suddenly stop moving.

All enemy scripts are now moved to EnemyScripts folder