Phase II: Core Programming - Feature: Aggresive Enemy Type
A new enemy, KamikaseBasicEnemy, will direct itself to the player when the player gets at certain distance to it.

Also, the new enemy has a new animation and sprite. The source is: https://opengameart.org/content/ufdish-bleeds-game-art?fbclid=IwAR2xEuQo9RWdbTst8A9x1UMXBOMm41IWHqLyhwZs52hHQFOWZyTfci6ydQk

The new script ESeekerSystem must be put in a child object from the Enemy, it will search for the EnemyCore and EnemyMovement components. If it detects that the player is close enough, it will take control of the movement and turn off movement in the EnemyMovement script. This seeker system can be added to any other enemy with EnemyCore and EnemyMovement.

Fixes:
- SpawnManager looks for EnemyCore components instead of Enemy components in order to know when the next wave should start.
- WaveDirector is now aware of the new enemy and adds it to the waves.