Phase II: Core Programming - Feature: Smart Enemy
A new system was created, BackWatcher, which makes the enemy shoots a laser backwards if the player goes behind it.

The BW system takes reference of the script that make able to shoot. It also needs to be applying the interface IWeapon, which defines Shoot(Vector2 direction) function. As longo as the weapon  used by the Enemy aplies IWeapon, the BackWatcher can be added to that enemy.

A new Laser Script was created to behave like LaserEnemy, but implementing IWeapon and giving direction to the laser. It is called LaserEnemyD.

Also, a new prefab, BasicDEnemy was created to support this new laser and the BackWatcher system.