Phase II: Core Programming - Feature: EnemyPickups.
Now enemies have the capability of destroying powerups (EnemyLaserDamage).

EnemyItemShooter: A new enemy was created, which additionaly to the common enemy behaviour, will increase shooting when it detects a powerup in front of it with the objective to destroy it.

This new behaviour is carried out by the ItemDestroyer component in the EnemySystems. All it needs is a IWeapon component to order it to shoot when it detects a powerup at the front.

This enemy has been added to the WaveDirector, and is slightly yellow to differentiate it from the other basic enemies.