Phase II: Core Programming - Feature: Enemy Avoid Shot
A new system has been created: the LaserDodger. This component searches for the EnemyMovement component. When is not dodging, the LD searches all available player's lasers, then selects the nearest to the EnemyShip and checks if is near enough (radar distance). If is near enough, it will check the side where it comes from and will set a new course to the down-diagonal of the oposite side. This new course is performed by the EnemyMovement component, which was modified to accept new directions from other components with two new functions: SetCustomDirection and RemoveCustomDirection.

When SetCustomDirection is used, EnemyMovement will follow that direction, and will ignore the default movement generator. When RemoveCustomDirection is used, it will remove the used custom direction and will start using the MovementGenerator again.

Once the laser watched by the LaserDodger is far enough, it will start searching again for another laser.

The new enemy with this component is EnemyLDodger, and has been added to the list of the WaveCreator, and included by the WaveDirector.