Phase II: Core Programming - Feature: Homming Projectile.
A new weapon is available for the player. When the Missile porwerup is picked up, the player can shot missiles that will target an enemy after reaching to certain distance to it. It will lock on that enemy and will redirect the missile to it. The missile has rotation velocity and translation velocity to make its movement more natural. If the missile lost track of the enemy locked (the enemy died or got out of its range) it will redirect to the top side of the screen, and will search on its way a new enemy.

This powerup is as rare as the energy shot. It has been added to the wave director.

Added new sprites for missile and powerup, also added a particle system for the missile thruster fire.